+++ src/types/set.rs
@@ -6,6 +6,7 @@ use crate::ffi;
 use crate::instance::PyNativeType;
 use crate::internal_tricks::Unsendable;
 use crate::object::PyObject;
+use crate::types::PyAny;
 use crate::AsPyPointer;
 use crate::Python;
 use crate::{ToBorrowedObject, ToPyObject};
@@ -97,7 +98,57 @@ impl PySet {
 
     /// Remove and return an arbitrary element from the set
     pub fn pop(&self) -> Option<PyObject> {
-        unsafe { PyObject::from_owned_ptr_or_opt(self.py(), ffi::PySet_Pop(self.as_ptr())) }
+        let element =
+            unsafe { PyObject::from_owned_ptr_or_err(self.py(), ffi::PySet_Pop(self.as_ptr())) };
+        match element {
+            Ok(e) => Some(e),
+            Err(_) => None,
+        }
+    }
+
+    /// Returns an iterator of values in this set.
+    ///
+    /// Note that it can be unsafe to use when the set might be changed by other code.
+    #[cfg(not(Py_LIMITED_API))]
+    pub fn iter(&self) -> PySetIterator {
+        PySetIterator {
+            set: self.as_ref(),
+            pos: 0,
+        }
+    }
+}
+
+#[cfg(not(Py_LIMITED_API))]
+pub struct PySetIterator<'py> {
+    set: &'py super::PyAny,
+    pos: isize,
+}
+
+#[cfg(not(Py_LIMITED_API))]
+impl<'py> Iterator for PySetIterator<'py> {
+    type Item = &'py super::PyAny;
+
+    #[inline]
+    fn next(&mut self) -> Option<Self::Item> {
+        unsafe {
+            let mut key: *mut ffi::PyObject = std::ptr::null_mut();
+            let mut hash: ffi::Py_hash_t = 0;
+            if ffi::_PySet_NextEntry(self.set.as_ptr(), &mut self.pos, &mut key, &mut hash) != 0 {
+                Some(self.set.py().from_borrowed_ptr(key))
+            } else {
+                None
+            }
+        }
+    }
+}
+
+#[cfg(not(Py_LIMITED_API))]
+impl<'a> std::iter::IntoIterator for &'a PySet {
+    type Item = &'a PyAny;
+    type IntoIter = PySetIterator<'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        self.iter()
     }
 }
 
@@ -171,15 +222,34 @@ impl PyFrozenSet {
             }
         })
     }
+
+    /// Returns an iterator of values in this frozen set.
+    ///
+    /// Note that it can be unsafe to use when the set might be changed by other code.
+    #[cfg(not(Py_LIMITED_API))]
+    pub fn iter(&self) -> PySetIterator {
+        self.into_iter()
+    }
+}
+
+#[cfg(not(Py_LIMITED_API))]
+impl<'a> std::iter::IntoIterator for &'a PyFrozenSet {
+    type Item = &'a PyAny;
+    type IntoIter = PySetIterator<'a>;
+
+    fn into_iter(self) -> Self::IntoIter {
+        PySetIterator {
+            set: self.as_ref(),
+            pos: 0,
+        }
+    }
 }
 
 #[cfg(test)]
 mod test {
     use super::{PyFrozenSet, PySet};
     use crate::instance::AsPyRef;
-    use crate::objectprotocol::ObjectProtocol;
-    use crate::Python;
-    use crate::{PyTryFrom, ToPyObject};
+    use crate::{ObjectProtocol, PyTryFrom, Python, ToPyObject};
     use std::collections::HashSet;
 
     #[test]
@@ -264,6 +334,9 @@ mod test {
         assert!(val.is_some());
         let val2 = set.pop();
         assert!(val2.is_none());
+        assert!(py
+            .eval("print('Exception state should not be set.')", None, None)
+            .is_ok());
     }
 
     #[test]
@@ -272,8 +345,15 @@ mod test {
         let py = gil.python();
 
         let set = PySet::new(py, &[1]).unwrap();
-        for el in set.iter().unwrap() {
-            assert_eq!(1i32, el.unwrap().extract::<i32>().unwrap());
+
+        // iter method
+        for el in set.iter() {
+            assert_eq!(1i32, el.extract().unwrap());
+        }
+
+        // intoiterator iteration
+        for el in set {
+            assert_eq!(1i32, el.extract().unwrap());
         }
     }
 
@@ -311,8 +391,15 @@ mod test {
         let py = gil.python();
 
         let set = PyFrozenSet::new(py, &[1]).unwrap();
-        for el in set.iter().unwrap() {
-            assert_eq!(1i32, el.unwrap().extract::<i32>().unwrap());
+
+        // iter method
+        for el in set.iter() {
+            assert_eq!(1i32, el.extract::<i32>().unwrap());
+        }
+
+        // intoiterator iteration
+        for el in set {
+            assert_eq!(1i32, el.extract::<i32>().unwrap());
         }
     }
 }